requirements.txt
numpy
# scipy
matplotlib
tqdm
# lpips
# mmcv
imageio
# imageio-ffmpeg
opencv-python
# torch_efficient_distloss
# Ninja
open3d
# tensorboard
# tensorboardX
termcolor
# pathlib3x
# pymcubes
scikit-image
# ipympl
icecream
trimesh
numpy-quaternion
# sympy
h5py
sympy
pyquaternion
ipdb
# openexr
jupyterlab
mitsuba
pytorch_lightning
bpy
pyhocon
cachetools
imageio
imagecodecs # if fails, try: conda install -c conda-forge imagecodecs
rawpy # https://github.com/letmaik/rawpy#installation-from-source-on-linuxmacos
termcolor

# === Also:
# conda install -c conda-forge ninja
